refactor: move to extension
